Build works again
Now uses webpack, wasm-pack, and wasm-bindgen
