+++ .gitignore
@@ -1,3 +1,4 @@
+/node_modules/
 /target/
 **/*.rs.bk
 Cargo.lock
