+++ js/index.js
@@ -1,0 +1,1 @@
+import("../pkg/index.js").catch(console.error);
