+++ src/rng.rs
@@ -1,7 +1,9 @@
 extern crate rand;
 
 pub use self::rand::{Rng, SeedableRng, StdRng};
+use wasm_bindgen::prelude::*;
 
+#[wasm_bindgen(module = "/js/demo.js")]
 extern "C" {
     fn js_fill_rand(ptr: *mut u8, len: usize) -> usize;
 }
@@ -14,13 +16,17 @@ pub enum RNGSourceError {
 }
 
 pub fn fill_random(buf: &mut [u8]) -> Result<(), RNGSourceError> {
-    let rv = unsafe { js_fill_rand(buf.as_mut_ptr(), buf.len()) };
-    match rv {
-        0 => Ok(()),
-        1 => Err(RNGSourceError::RangeError),
-        2 => Err(RNGSourceError::QuotaError),
-        _ => Err(RNGSourceError::UnknownError),
+    for (i, b) in buf.iter_mut().enumerate() {
+        *b = i as u8;
     }
+    //let rv = unsafe { js_fill_rand(buf.as_mut_ptr(), buf.len()) };
+    // match rv {
+    //     0 => Ok(()),
+    //     1 => Err(RNGSourceError::RangeError),
+    //     2 => Err(RNGSourceError::QuotaError),
+    //     _ => Err(RNGSourceError::UnknownError),
+    // }
+    Ok(())
 }
 
 pub fn new_rng() -> Result<StdRng, RNGSourceError> {
