+++ static/index.html
@@ -1,0 +1,15 @@
+<style>
+  svg {
+    background: black;
+  }
+  #path {
+    stroke-width: 2.5px;
+    stroke: white;
+  }
+</style>
+<svg width="1280" height="820">
+  <g transform="translate(0, 100)">
+    <path id="path"></path>
+  </g>
+</svg>
+<script type="module" src="index.js"></script>
